p34: lower bound a bit.
Thinking: More than 7 nines is impossible because, for example, 8 * 9! < 1e7.  So seven nines is a simple upper bound.

However, 7 nines only gets you to a 2 on the left-most digit.  So the first digit can't be a nine.  So six 9s with a 2 in the highest place gives a bound.
But there are only two possibilities there, and neither one matches.

In addition, the last digit can only be a nine if there are three other digits that are non-nine.  So five 9s is the most possible nines.  5 * 9! is 1814400.  To get the 5 middle digits up to 9s with the two remaining digits is impossible, even if they could both be eights, which they can't.  So we use 4 * 9! + 3 * 8! as an upper bound.  With more work we can do better, of course.